+++ wled00/xml.cpp
@@ -236,7 +236,7 @@ void getSettingsJS(byte subPage, char* dest)
     sappend('c',SET_F("WS"),noWifiSleep);
 
     #ifdef WLED_USE_ETHERNET
-    sappend('i',SET_F("ETH"),ethernetType);
+    sappend('v',SET_F("ETH"),ethernetType);
     #else
     //hide ethernet setting if not compiled in
     oappend(SET_F("document.getElementById('ethd').style.display='none';"));
@@ -270,10 +270,10 @@ void getSettingsJS(byte subPage, char* dest)
   }
 
   if (subPage == 2) {
-    char nS[3];
+    char nS[8];
 
     // add usermod pins as d.um_p array (TODO: usermod config shouldn't use state. instead we should load "um" object from cfg.json)
-    DynamicJsonDocument doc(JSON_BUFFER_SIZE);
+    /*DynamicJsonDocument doc(JSON_BUFFER_SIZE);
     JsonObject mods = doc.createNestedObject(F("mods"));
     usermods.addToJsonState(mods);
     if (!mods.isNull()) {
@@ -286,20 +286,18 @@ void getSettingsJS(byte subPage, char* dest)
         }
       }
       oappend(SET_F("];"));
-    }
+    }*/
 
-    #if defined(WLED_MAX_BUSSES) && WLED_MAX_BUSSES>1
-      oappend(SET_F("addLEDs("));
-      oappend(itoa(WLED_MAX_BUSSES,nS,10));
-      oappend(SET_F(");"));
-    #endif
+    oappend(SET_F("bLimits("));
+    oappend(itoa(WLED_MAX_BUSSES,nS,10));
+    oappend(",");
+    oappend(itoa(MAX_LEDS_PER_BUS,nS,10));
+    oappend(",");
+    oappend(itoa(MAX_LED_MEMORY,nS,10));
+    oappend(SET_F(");"));
 
     oappend(SET_F("d.Sf.LC.max=")); //TODO Formula for max LEDs on ESP8266 depending on types. 500 DMA or 1500 UART (about 4kB mem usage)
-    #if defined(ESP8266) && LEDPIN == 3
-    oappendi(MAX_LEDS_DMA);
-    #else
     oappendi(MAX_LEDS);
-    #endif
     oappend(";");
 
     sappend('v',SET_F("LC"),ledCount);
@@ -307,7 +305,6 @@ void getSettingsJS(byte subPage, char* dest)
     for (uint8_t s=0; s < busses.getNumBusses(); s++){
       Bus* bus = busses.getBus(s);
       char lp[4] = "L0"; lp[2] = 48+s; lp[3] = 0; //ascii 0-9 //strip data pin
-      char lk[4] = "L1"; lk[2] = 48+s; lk[3] = 0; //strip clock pin. 255 for none
       char lc[4] = "LC"; lc[2] = 48+s; lc[3] = 0; //strip length
       char co[4] = "CO"; co[2] = 48+s; co[3] = 0; //strip color order
       char lt[4] = "LT"; lt[2] = 48+s; lt[3] = 0; //strip type
@@ -316,8 +313,10 @@ void getSettingsJS(byte subPage, char* dest)
       oappend(SET_F("addLEDs(1);"));
       uint8_t pins[5];
       uint8_t nPins = bus->getPins(pins);
-      sappend('v', lp, pins[0]);
-      if (pinManager.isPinOk(pins[1])) sappend('v', lk, pins[1]);
+      for (uint8_t i = 0; i < nPins; i++) {
+        lp[1] = 48+i;
+        if (pinManager.isPinOk(pins[i])) sappend('v', lp, pins[i]);
+      }
       sappend('v', lc, bus->getLength());
       sappend('v',lt,bus->getType());
       sappend('v',co,bus->getColorOrder());
@@ -328,7 +327,7 @@ void getSettingsJS(byte subPage, char* dest)
     sappend('v',SET_F("LA"),strip.milliampsPerLed);
     if (strip.currentMilliamps)
     {
-      sappends('m',SET_F("(\"pow\")[0]"),"");
+      sappends('m',SET_F("(\"pow\")[0]"),(char*)"");
       olen -= 2; //delete ";
       oappendi(strip.currentMilliamps);
       oappend(SET_F("mA\";"));