+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 ### Development versions after 0.11.1 release
 
+#### Build 2103090
+
+-   Added Auto Node discovery (PR #1683)
+-   Added tooltips to quick color selectors for accessibility
+
+#### Build 2103060
+
+-   Auto start field population in bus config
+
+#### Build 2103050
+
+-   Fixed incorrect over-memory indication in LED settings on ESP32
+
+#### Build 2103041
+
+-   Added destructor for BusPwm (fixes #1789)
+
+#### Build 2103040
+
+-   Fixed relay mode inverted when upgrading from 0.11.0
+-   Fixed no more than 2 pins per bus configurable in UI
+-   Changed to non-linear IR brightness steps (PR #1742)
+-   Fixed various warnings (PR #1744)
+-   Added UDP DNRGBW Mode (PR #1704)
+-   Added dynamic LED mapping with ledmap.json file (PR #1738)
+-   Added support for QuinLED-ESP32-Ethernet board
+-   Added support for WESP32 ethernet board (PR #1764)
+-   Added Caching for main UI (PR #1704)
+-   Added Tetrix mode (PR #1729)
+-   Added memory check on Bus creation
+
 #### Build 2102050
 
 -   Version bump to 0.12.0-a0 "Hikari"
@@ -73,7 +104,7 @@
 
 -   Added Preset ID quick display option (PR #1462)
 -   Fixed LEDs not turning on when using gamma correct brightness and LEDPIN 2 (default)
--   Fixed notifier applying main segment to selected segments on notification with FX/Col disabled 
+-   Fixed notifier applying main segment to selected segments on notification with FX/Col disabled
 
 #### Build 2012130
 
@@ -149,7 +180,7 @@
 -   Added HEX color receiving to JSON API with `"col":["RRGGBBWW"]` format
 -   Moved Kelvin color receiving in JSON API from `"col":[[val]]` to `"col":[val]` format
     _Notice:_ This is technically a breaking change. Since no release was made since the introduction and the Kelvin property was not previously documented in the wiki,
-    impact should be minimal. 
+    impact should be minimal.
 -   BTNPIN can now be disabled by setting to -1 (fixes #1237)
 
 #### Build 2011180
@@ -225,7 +256,7 @@
 #### Build 2010020
 
 -   Fixed interaction of `T` and `NL` HTTP API commands (#1214)
--   Fixed an issue where Sunrise mode nightlight does not activate if toggled on simultaneously 
+-   Fixed an issue where Sunrise mode nightlight does not activate if toggled on simultaneously
 
 #### Build 2009291